feat: Content-cache-based isUploaded judgment, retire is_unsynced
